+++ types/selector-id.d.ts
@@ -2,4 +2,4 @@ import { CssSelector } from './types';
 /**
  * Get ID selector for an element.
  */
-export declare function getIdSelector(element: Element): Array<CssSelector>;
+export declare function getIdSelector(elements: Element[]): Array<CssSelector>;