+++ types/selector-tag.d.ts
@@ -2,4 +2,4 @@ import { CssSelector } from './types';
 /**
  * Get tag selector for an element.
  */
-export declare function getTagSelector(element: Element): Array<CssSelector>;
+export declare function getTagSelector(elements: Element[]): Array<CssSelector>;