+++ types/utilities-dom.d.ts
@@ -2,11 +2,19 @@ import { CssSelector } from './types';
 /**
  * Check whether element is matched uniquely by selector.
  */
-export declare function testSelector(element: Element, selector: CssSelector, root?: ParentNode): boolean;
+export declare function testSelector(elements: Element[], selector: CssSelector, root?: ParentNode): boolean;
 /**
- * Find all parent elements of the element.
+ * Test whether selector targets element. It does not have to be a unique match.
  */
-export declare function getParents(element: Element, root?: ParentNode): Array<Element>;
+export declare function testMultiSelector(element: Element, selector: CssSelector, root?: ParentNode): boolean;
+/**
+ * Find all parents of a single element.
+ */
+export declare function getElementParents(element: Element, root: ParentNode): Element[];
+/**
+ * Find all common parents of elements.
+ */
+export declare function getParents(elements: Element[], root?: ParentNode): Element[];
 /**
  * Returns root node for given element. This needs to be used because of document-less environments, e.g. jsdom.
  */